Moved StyledTextAreaView into its own source file.
